+++ storage/innobase/btr/btr0btr.cc
@@ -1777,7 +1777,7 @@ void btr_set_instant(buf_block_t* root, const dict_index_t& index, mtr_t* mtr)
 		ut_ad(!memcmp(supremum, "supremum", 8));
 		mtr->write<2>(*root, page_type, FIL_PAGE_TYPE_INSTANT);
 		ut_ad(i <= PAGE_NO_DIRECTION);
-		i |= index.n_core_fields << 3;
+		i |= static_cast<uint16_t>(index.n_core_fields << 3);
 		mtr->write<2>(*root, PAGE_HEADER + PAGE_INSTANT + root->frame,
 			      i);
 		break;
